ref(parsing_mod): restructure parsing module
