+++ src/event_parsing/attachment.rs
@@ -2,12 +2,12 @@ use std::convert::TryFrom;
 
 use nom::{bytes::complete::take, combinator::map, error::ErrorKind, multi::count};
 
-use crate::{derivation::attached_signature_code::b64_to_num, event::sections::seal::{EventSeal, SourceSeal}, event_message::parse::Attachment, event_parsing::payload_size::PayloadType, prefix::{AttachedSignaturePrefix, BasicPrefix, SelfSigningPrefix}};
+use crate::{derivation::attached_signature_code::b64_to_num, event::sections::seal::{EventSeal, SourceSeal}, event_parsing::payload_size::PayloadType, prefix::{AttachedSignaturePrefix, BasicPrefix, SelfSigningPrefix}};
 
-use super::prefix::{
+use super::{Attachment, prefix::{
     attached_signature, attached_sn, basic_prefix, prefix, self_addressing_prefix,
     self_signing_prefix,
-};
+}};
 
 /// returns attached source seals
 fn source_seal(s: &[u8]) -> nom::IResult<&[u8], Vec<SourceSeal>> {
@@ -33,7 +33,7 @@ fn event_seal(s: &[u8]) -> nom::IResult<&[u8], EventSeal> {
     let (rest, event_digest) = self_addressing_prefix(rest)?;
     let seal = EventSeal {
         prefix: identifier,
-        sn: u64::from(sn),
+        sn,
         event_digest,
     };
