+++ src/event_parsing/mod.rs
@@ -1,170 +1,350 @@
-use std::io::Cursor;
-
-use nom::{
-    branch::alt,
-    error::ErrorKind,
-    multi::{fold_many0, many0},
-};
+use std::convert::TryFrom;
+use base64::URL_SAFE_NO_PAD;
 use serde::Deserialize;
 
-#[cfg(feature = "async")]
-use crate::event_message::serialization_info::SerializationInfo;
+use crate::event::EventMessage;
+use crate::event::sections::seal::{EventSeal, SourceSeal};
+use crate::event_message::signed_event_message::{Message, SignedEventMessage, SignedNontransferableReceipt, SignedTransferableReceipt};
+use crate::event_parsing::payload_size::PayloadType;
+use crate::prefix::{AttachedSignaturePrefix, BasicPrefix, Prefix, SelfSigningPrefix};
+use crate::{error::Error, event::event_data::EventData};
 
-use crate::{
-    event::EventMessage,
-    event_message::{parse::Attachment, parse::DeserializedSignedEvent},
-};
-use rmp_serde as serde_mgpk;
 pub mod attachment;
 pub mod payload_size;
 pub mod prefix;
+pub mod message;
 
-fn json_message(s: &[u8]) -> nom::IResult<&[u8], EventMessage> {
-    let mut stream = serde_json::Deserializer::from_slice(s).into_iter::<EventMessage>();
-    match stream.next() {
-        Some(Ok(event)) => Ok((&s[stream.byte_offset()..], event)),
-        _ => Err(nom::Err::Error((s, ErrorKind::IsNot))),
+#[derive(Debug, Clone, Deserialize, PartialEq)]
+pub enum Attachment {
+    SealSourceCouplets(Vec<SourceSeal>),
+    AttachedEventSeal(Vec<EventSeal>),
+    AttachedSignatures(Vec<AttachedSignaturePrefix>),
+    ReceiptCouplets(Vec<(BasicPrefix, SelfSigningPrefix)>),
+}
+
+impl Attachment {
+    pub fn to_cesr(&self) -> String {
+         let (payload_type, att_len, serialized_attachment) = match self {
+            Attachment::SealSourceCouplets(sources) => {
+                let serialzied_sources = sources
+                    .iter()
+                    .fold("".into(), |acc, s| [acc, Self::pack_sn(s.sn), s.digest.to_str()].join(""));
+
+                (PayloadType::MG, sources.len(), serialzied_sources)
+            }
+            Attachment::AttachedEventSeal(seal) => {
+                let serialized_seals = seal.iter().fold("".into(), |acc, seal| {
+                    [
+                        acc,
+                        seal.prefix.to_str(),
+                        Self::pack_sn(seal.sn),
+                        seal.event_digest.to_str(),
+                    ]
+                    .join("")
+                });
+                (PayloadType::MF, seal.len(), serialized_seals)
+            }
+            Attachment::AttachedSignatures(sigs) => {
+                let serialized_sigs = sigs
+                    .iter()
+                    .fold("".into(), |acc, sig| [acc, sig.to_str()].join(""));
+                (PayloadType::MA, sigs.len(), serialized_sigs)
+            }
+            Attachment::ReceiptCouplets(couplets) => {
+                let packed_couplets = couplets.iter().fold("".into(), |acc, (bp, sp)| {
+                    [acc, bp.to_str(), sp.to_str()].join("")
+                });
+
+                (PayloadType::MC, couplets.len(), packed_couplets)
+            }
+        };
+        [
+            payload_type.adjust_with_num(att_len as u16),
+            serialized_attachment,
+        ]
+        .join("")
     }
-}
 
-fn cbor_message(s: &[u8]) -> nom::IResult<&[u8], EventMessage> {
-    let mut stream = serde_cbor::Deserializer::from_slice(s).into_iter::<EventMessage>();
-    match stream.next() {
-        Some(Ok(event)) => Ok((&s[stream.byte_offset()..], event)),
-        _ => Err(nom::Err::Error((s, ErrorKind::IsNot))),
+    fn pack_sn(sn: u64) -> String {
+        let payload_type = PayloadType::OA;
+        let sn_raw: Vec<u8> = sn.to_be_bytes().into();
+        // Calculate how many zeros are missing to achieve expected base64 string
+        // length. Master code size is expected padding size.
+        let missing_zeros =
+            payload_type.size() / 4 * 3 - payload_type.master_code_size(false) - sn_raw.len();
+        let sn_vec: Vec<u8> = std::iter::repeat(0)
+                .take(missing_zeros)
+                .chain(sn_raw)
+                .collect();
+            [
+                payload_type.to_string(),
+                base64::encode_config(sn_vec, URL_SAFE_NO_PAD),
+            ]
+            .join("")
+        }
     }
+
+#[derive(Clone, Debug, PartialEq)]
+pub struct SignedEventData {
+    pub deserialized_event: EventMessage,
+    pub attachments: Vec<Attachment>,
 }
 
-fn mgpk_message(s: &[u8]) -> nom::IResult<&[u8], EventMessage> {
-    let mut deser = serde_mgpk::Deserializer::new(Cursor::new(s));
-    match Deserialize::deserialize(&mut deser) {
-        Ok(event) => Ok((&s[deser.get_ref().position() as usize..], event)),
-        _ => Err(nom::Err::Error((s, ErrorKind::IsNot))),
-    }
-}
-
-pub fn message<'a>(s: &'a [u8]) -> nom::IResult<&[u8], EventMessage> {
-    alt((json_message, cbor_message, mgpk_message))(s).map(|d| (d.0, d.1))
-}
-
-pub fn signed_message<'a>(s: &'a [u8]) -> nom::IResult<&[u8], DeserializedSignedEvent> {
-    let (rest, msg) = message(s)?;
-    let (rest, attachments): (&[u8], Vec<Attachment>) =
-        fold_many0(attachment::attachment, vec![], |mut acc: Vec<_>, item| {
-            acc.push(item);
-            acc
-        })(rest)?;
-
-    Ok((
-        rest,
-        DeserializedSignedEvent {
-            deserialized_event: msg,
+impl SignedEventData {
+    pub fn to_cesr(&self) -> Result<Vec<u8>, Error> {
+        let attachments = self.attachments
+            .iter()
+            .fold(String::default(), |acc, att| [acc, att.to_cesr()].concat())
+            .as_bytes().to_vec();
+        Ok([
+            self.deserialized_event.serialize()?,
             attachments,
-        },
-    ))
+        ]
+        .concat())
+    }    
 }
 
-pub fn signed_event_stream(s: &[u8]) -> nom::IResult<&[u8], Vec<DeserializedSignedEvent>> {
-    many0(signed_message)(s)
+impl From<&SignedEventMessage> for SignedEventData {
+    fn from(ev: &SignedEventMessage) -> Self {
+        let attachments = match ev.delegator_seal.clone() {
+            Some(delegator_seal) => 
+                [
+                    Attachment::SealSourceCouplets(vec![delegator_seal]),
+                    Attachment::AttachedSignatures(ev.signatures.clone()) 
+                ].into(),
+            None => [Attachment::AttachedSignatures(ev.signatures.clone())].into(),
+        }; 
+        
+        SignedEventData { deserialized_event: ev.event_message.clone(), attachments }
+    }
+    
 }
 
-// TESTED: OK
-#[cfg(feature = "async")]
-fn json_version(data: &[u8]) -> nom::IResult<&[u8], SerializationInfo> {
-    match serde_json::from_slice(data) {
-        Ok(vi) => Ok((data, vi)),
-        _ => Err(nom::Err::Error((data, ErrorKind::IsNot))),
+impl From<SignedNontransferableReceipt> for SignedEventData {
+    fn from(rcp: SignedNontransferableReceipt) -> SignedEventData {
+        let attachments = [Attachment::ReceiptCouplets(rcp.couplets)].into();
+        SignedEventData { deserialized_event: rcp.body, attachments }
     }
 }
 
-// TODO: Requires testing
-#[cfg(feature = "async")]
-fn cbor_version(data: &[u8]) -> nom::IResult<&[u8], SerializationInfo> {
-    match serde_cbor::from_slice(data) {
-        Ok(vi) => Ok((data, vi)),
-        _ => Err(nom::Err::Error((data, ErrorKind::IsNot))),
+impl From<SignedTransferableReceipt> for SignedEventData {
+    fn from(rcp: SignedTransferableReceipt) -> SignedEventData {
+        let attachments = [
+                Attachment::AttachedEventSeal(vec![rcp.validator_seal]), 
+                Attachment::AttachedSignatures(rcp.signatures)
+            ].into();
+        SignedEventData { deserialized_event: rcp.body, attachments }
     }
 }
 
-// TODO: Requires testing
-#[cfg(feature = "async")]
-fn mgpk_version(data: &[u8]) -> nom::IResult<&[u8], SerializationInfo> {
-    match serde_mgpk::from_slice(data) {
-        Ok(vi) => Ok((data, vi)),
-        _ => Err(nom::Err::Error((data, ErrorKind::IsNot))),
+impl TryFrom<SignedEventData> for Message {
+    type Error = Error;
+
+    fn try_from(value: SignedEventData) -> Result<Self, Self::Error> {
+        signed_message(value)
     }
 }
 
-#[cfg(feature = "async")]
-pub(crate) fn version<'a>(data: &'a [u8]) -> nom::IResult<&[u8], SerializationInfo> {
-    alt((json_version, cbor_version, mgpk_version))(data).map(|d| (d.0, d.1))
+fn signed_message(mut des: SignedEventData) -> Result<Message, Error> {
+    match des.deserialized_event.event.event_data {
+        EventData::Rct(_) => {
+            let att = des.attachments.pop().ok_or_else(|| Error::SemanticError("Missing attachment".into()))?;
+            match att {
+                // Should be nontransferable receipt
+                Attachment::ReceiptCouplets(couplets) => 
+                Ok(
+                    Message::NontransferableRct(SignedNontransferableReceipt {
+                        body: des.deserialized_event,
+                        couplets,
+                    })
+                ),
+                Attachment::AttachedEventSeal(_) | Attachment::AttachedSignatures(_) => {
+                    // Should be transferable receipt
+                    let second_att = des.attachments.pop().ok_or_else(|| Error::SemanticError("Missing attachment".into()))?;
+
+                    let (seals, sigs) = match (att, second_att) {
+                        (
+                            Attachment::AttachedEventSeal(seals),
+                            Attachment::AttachedSignatures(sigs),
+                        ) => Ok((seals, sigs)),
+                        (
+                            Attachment::AttachedSignatures(sigs),
+                            Attachment::AttachedEventSeal(seals),
+                        ) => Ok((seals, sigs)),
+                        _ => {
+                            // improper attachments
+                            Err(Error::SemanticError("Improper attachment".into()))
+                        }
+                    }?;
+
+                    Ok(
+                        Message::TransferableRct(SignedTransferableReceipt::new(
+                            &des.deserialized_event,
+                            // TODO what if more than one?
+                            seals
+                                .last()
+                                .ok_or_else(|| Error::SemanticError("More than one seal".into()))?
+                                .to_owned(),
+                            sigs,
+                        )),
+                    )
+                }
+                _ => {
+                    // Improper payload type
+                    Err(Error::SemanticError("Improper payload type".into()))
+                }
+            }
+        }
+        EventData::Dip(_) | EventData::Drt(_) => {
+            let (att1, att2) = (
+                des.attachments.pop().ok_or_else(|| Error::SemanticError("Missing attachment".into()))?, 
+                des.attachments.pop().ok_or_else(|| Error::SemanticError("Missing attachment".into()))?,
+            );
+
+            let (seals, sigs) = match (att1, att2) {
+                (Attachment::SealSourceCouplets(seals), Attachment::AttachedSignatures(sigs)) => {
+                    Ok((seals, sigs))
+                }
+                (Attachment::AttachedSignatures(sigs), Attachment::SealSourceCouplets(seals)) => {
+                    Ok((seals, sigs))
+                }
+                _ => {
+                    // Improper attachment type
+                    Err(Error::SemanticError("Improper attachment type".into()))
+                }
+            }?;
+            let delegator_seal = match seals.len() {
+                0 => Err(Error::SemanticError("Missing delegator seal".into())),
+                1 => Ok(seals.first().cloned()),
+                _ => Err(Error::SemanticError("Too many seals".into())),
+            };
+            
+            Ok(
+                Message::Event(
+                    SignedEventMessage::new(&des.deserialized_event, sigs, delegator_seal?)
+                ),
+            )
+        }
+        _ => {
+            let sigs = des.attachments.first().cloned().ok_or_else(|| Error::SemanticError("Missing attachment".into()))?;
+            if let Attachment::AttachedSignatures(sigs) = sigs {
+                Ok(
+                    Message::Event(
+                    SignedEventMessage::new(&des.deserialized_event, sigs.to_vec(), None)
+                ))
+            } else {
+                // Improper attachment type
+                Err(Error::SemanticError("Improper attachment type".into()))
+            }
+        }
+    }
 }
 
-#[cfg(feature = "async")]
+// pub fn signed_event_stream_validate(s: &[u8]) -> nom::IResult<&[u8], IdentifierState> {
+//     let (rest, id) = fold_many1(
+//         signed_message,
+//         Ok(IdentifierState::default()),
+//         |acc, next| match next {
+//             Deserialized::Event(e) => {
+//                 let new_state = acc?
+//                     .apply(&e.deserialized_event)
+//                     .map_err(|_| nom::Err::Error((s, ErrorKind::Verify)))?;
+//                 if new_state
+//                     .current
+//                     .verify(&e.deserialized_event.serialize().unwrap(), &e.signatures)
+//                     .map_err(|_| nom::Err::Error((s, ErrorKind::Verify)))?
+//                 {
+//                     Ok(new_state)
+//                 } else {
+//                     Err(nom::Err::Error((s, ErrorKind::Verify)))
+//                 }
+//             }
+//             // TODO this probably should not just skip non-events
+//             _ => acc,
+//         },
+//     )(s)?;
+
+//     Ok((rest, id?))
+// }
+
 #[test]
-fn test_version_parse() {
-    let json = br#""KERI10JSON00014b_""#;
-    let json_result = version(json);
-    assert!(json_result.is_ok());
+fn test_stream1() {
+    use crate::event_parsing;
+    // taken from KERIPY: tests/core/test_eventing.py::test_kevery#1998
+    let stream = br#"{"v":"KERI10JSON0000ed_","i":"DSuhyBcPZEZLK-fcw5tzHn2N46wRCG_ZOoeKtWTOunRA","s":"0","t":"icp","kt":"1","k":["DSuhyBcPZEZLK-fcw5tzHn2N46wRCG_ZOoeKtWTOunRA"],"n":"EPYuj8mq_PYYsoBKkzX1kxSPGYBWaIya3slgCOyOtlqU","bt":"0","b":[],"c":[],"a":[]}-AABAAmagesCSY8QhYYHCJXEWpsGD62qoLt2uyT0_Mq5lZPR88JyS5UrwFKFdcjPqyKc_SKaKDJhkGWCk07k_kVkjyCA"#;
+
+    let parsed = event_parsing::message::signed_message(stream).unwrap().1;
+    let msg= signed_message(parsed).unwrap();
+    assert!(matches!(msg, Message::Event(_)));
+
+    match msg {
+        Message::Event(signed_event) => {
+            assert_eq!(
+                signed_event.event_message.serialize().unwrap().len(),
+                signed_event
+                    .event_message
+                    .serialization_info
+                    .size
+            );
+
+            let serialized_again = signed_event.serialize();
+            assert!(serialized_again.is_ok());
+            let stringified = String::from_utf8(serialized_again.unwrap()).unwrap();
+            assert_eq!(stream, stringified.as_bytes())
+        }
+        _ => assert!(false),
+    }
 }
 
 #[test]
-fn test_signed_event() {
+fn test_stream2() {
+    use crate::event_parsing;
     // taken from KERIPY: tests/core/test_eventing.py::test_multisig_digprefix#2244
     let stream = br#"{"v":"KERI10JSON00014b_","i":"EsiHneigxgDopAidk_dmHuiUJR3kAaeqpgOAj9ZZd4q8","s":"0","t":"icp","kt":"2","k":["DSuhyBcPZEZLK-fcw5tzHn2N46wRCG_ZOoeKtWTOunRA","DVcuJOOJF1IE8svqEtrSuyQjGTd2HhfAkt9y2QkUtFJI","DT1iAhBWCkvChxNWsby2J0pJyxBIxbAtbLA0Ljx-Grh8"],"n":"E9izzBkXX76sqt0N-tfLzJeRqj0W56p4pDQ_ZqNCDpyw","bt":"0","b":[],"c":[],"a":[]}-AADAAhcaP-l0DkIKlJ87iIVcDx-m0iKPdSArEu63b-2cSEn9wXVGNpWw9nfwxodQ9G8J3q_Pm-AWfDwZGD9fobWuHBAAB6mz7zP0xFNBEBfSKG4mjpPbeOXktaIyX8mfsEa1A3Psf7eKxSrJ5Woj3iUB2AhhLg412-zkk795qxsK2xfdxBAACj5wdW-EyUJNgW0LHePQcSFNxW3ZyPregL4H2FoOrsPxLa3MZx6xYTh6i7YRMGY50ezEjV81hkI1Yce75M_bPCQ"#;
 
-    let parsed = signed_message(stream);
-    assert!(parsed.is_ok());
+    let parsed = event_parsing::message::signed_message(stream).unwrap().1;
+    let msg = signed_message(parsed);
+    assert!(msg.is_ok());
+    assert!(matches!(msg, Ok(Message::Event(_))));
+
+    match msg.unwrap() {
+        Message::Event(signed_event) => {
+            assert_eq!(
+                signed_event.event_message.serialize().unwrap().len(),
+                signed_event
+                    .event_message
+                    .serialization_info
+                    .size
+            );
+
+            let serialized_again = signed_event.serialize();
+            assert!(serialized_again.is_ok());
+            let stringified = String::from_utf8(serialized_again.unwrap()).unwrap();
+            assert_eq!(stream, stringified.as_bytes())
+        }
+        _ => assert!(false),
+    }
+
 }
 
 #[test]
-fn test_event() {
-    let stream = br#"{"v":"KERI10JSON0000ed_","i":"E7WIS0e4Tx1PcQW5Um5s3Mb8uPSzsyPODhByXzgvmAdQ","s":"0","t":"icp","kt":"1","k":["Dpt7mGZ3y5UmhT1NLExb1IW8vMJ8ylQW3K44LfkTgAqE"],"n":"Erpltchg7BUv21Qz3ZXhOhVu63m7S7YbPb21lSeGYd90","bt":"0","b":[],"c":[],"a":[]}"#;
-    let event = message(stream);
-    assert!(event.is_ok());
-    assert_eq!(event.unwrap().1.serialize().unwrap(), stream);
+fn test_deserialize() {
+    use crate::event_parsing::message::signed_message;
+    let trans_receipt_event = br#"{"v":"KERI10JSON000091_","i":"E7WIS0e4Tx1PcQW5Um5s3Mb8uPSzsyPODhByXzgvmAdQ","s":"0","t":"rct","d":"ErDNDBG7x2xYAH2i4AOnhVe44RS3lC1mRRdkyolFFHJk"}-FABENlofRlu2VPul-tjDObk6bTia2deG6NMqeFmsXhAgFvA0AAAAAAAAAAAAAAAAAAAAAAAE_MT0wsz-_ju_DVK_SaMaZT9ZE7pP4auQYeo2PDaw9FI-AABAA0Q7bqPvenjWXo_YIikMBKOg-pghLKwBi1Plm0PEqdv67L1_c6dq9bll7OFnoLp0a74Nw1cBGdjIPcu-yAllHAw"#;
+    let parsed_trans_receipt = signed_message(trans_receipt_event).unwrap().1;
+    let msg = Message::try_from(parsed_trans_receipt); 
+    assert!(matches!(msg, Ok(Message::TransferableRct(_))));
+    assert!(msg.is_ok());
 
-    // Inception event.
-    let stream = r#"{"v":"KERI10JSON00011c_","i":"EZAoTNZH3ULvaU6Z-i0d8JJR2nmwyYAfSVPzhzS6b5CM","s":"0","t":"icp","kt":"1","k":["DaU6JR2nmwyZ-i0d8JZAoTNZH3ULvYAfSVPzhzS6b5CM"],"n":"EZ-i0d8JZAoTNZH3ULvaU6JR2nmwyYAfSVPzhzS6b5CM","bt":"1","b":["DTNZH3ULvaU6JR2nmwyYAfSVPzhzS6bZ-i0d8JZAo5CM"],"c":["EO"],"a":[]}"#.as_bytes();
-    let event = message(stream);
-    assert!(event.is_ok());
-    assert_eq!(event.unwrap().1.serialize().unwrap(), stream);
+    // Taken from keripy/core/test_witness.py
+    let nontrans_rcp = br#"{"v":"KERI10JSON000091_","i":"EpU9D_puIW_QhgOf3WKUy-gXQnXeTQcJCO_Igcxi1YBg","s":"0","t":"rct","d":"EIt0xQQf-o-9E1B9VTDHiicQzVWk1CptvnewcnuhSd0M"}-CABB389hKezugU2LFKiFVbitoHAxXqJh6HQ8Rn9tH7fxd680BCZrTPLvG7sNaxtV8ZGdIHABFHCZ9FlnG6b4J6a9GcyzJIJOjuGNphW2zyC_WWU6CGMG7V52UeJxPqLpaYdP7Cg"#;
+    let parsed_nontrans_receipt = signed_message(nontrans_rcp).unwrap().1;
+    let msg = Message::try_from(parsed_nontrans_receipt);
+    assert!(msg.is_ok());
+    assert!(matches!(msg, Ok(Message::NontransferableRct(_))));
 
-    // Rotation event.
-    let stream = r#"{"v":"KERI10JSON00011c_","i":"EZAoTNZH3ULvaU6Z-i0d8JJR2nmwyYAfSVPzhzS6b5CM","s":"1","t":"rot","p":"EULvaU6JR2nmwyZ-i0d8JZAoTNZH3YAfSVPzhzS6b5CM","kt":"1","k":["DaU6JR2nmwyZ-i0d8JZAoTNZH3ULvYAfSVPzhzS6b5CM"],"n":"EYAfSVPzhzZ-i0d8JZAoTNZH3ULvaU6JR2nmwyS6b5CM","bt":"1","br":["DH3ULvaU6JR2nmwyYAfSVPzhzS6bZ-i0d8TNZJZAo5CM"],"ba":["DTNZH3ULvaU6JR2nmwyYAfSVPzhzS6bZ-i0d8JZAo5CM"],"a":[{"i":"EJJR2nmwyYAfSVPzhzS6b5CMZAoTNZH3ULvaU6Z-i0d8","s":"0","d":"ELvaU6Z-i0d8JJR2nmwyYAZAoTNZH3UfSVPzhzS6b5CM"}]}"#.as_bytes();
-    let event = message(stream);
-    assert!(event.is_ok());
-    assert_eq!(event.unwrap().1.serialize().unwrap(), stream);
-
-    // Interaction event without seals.
-    let stream = r#"{"v":"KERI10JSON0000a3_","i":"DSuhyBcPZEZLK-fcw5tzHn2N46wRCG_ZOoeKtWTOunRA","s":"3","t":"ixn","p":"EHBaMkc2lTj-1qnIgSeD0GmYjw8Zv6EmCgGDVPedn3fI","a":[]}"#.as_bytes();
-    let event = message(stream);
-    assert!(event.is_ok());
-    assert_eq!(event.unwrap().1.serialize().unwrap(), stream);
-
-    // Interaction event with seal.
-    let stream = r#"{"v":"KERI10JSON00011c_","i":"EZAoTNZH3ULvaU6Z-i0d8JJR2nmwyYAfSVPzhzS6b5CM","s":"2","t":"ixn","p":"EULvaU6JR2nmwyZ-i0d8JZAoTNZH3YAfSVPzhzS6b5CM","a":[{"i":"EJJR2nmwyYAfSVPzhzS6b5CMZAoTNZH3ULvaU6Z-i0d8","s":"1","d":"ELvaU6Z-i0d8JJR2nmwyYAZAoTNZH3UfSVPzhzS6b5CM"}]}"#.as_bytes();
-    let event = message(stream);
-    assert!(event.is_ok());
-    assert_eq!(event.unwrap().1.serialize().unwrap(), stream);
-
-    // Delegated inception event.
-    let stream = r#"{"v":"KERI10JSON000121_","i":"E-9tsnVcfUyXVQyBPGfntoL-xexf4Cldt_EPzHis2W4U","s":"0","t":"dip","kt":"1","k":["DuK1x8ydpucu3480Jpd1XBfjnCwb3dZ3x5b1CJmuUphA"],"n":"EWWkjZkZDXF74O2bOQ4H5hu4nXDlKg2m4CBEBkUxibiU","bt":"0","b":[],"c":[],"a":[],"di":"Eta8KLf1zrE5n-HZpgRAnDmxLASZdXEiU9u6aahqR8TI"}"#.as_bytes();
-    let event = message(stream);
-    assert_eq!(event.unwrap().1.serialize().unwrap(), stream);
-
-    // // Delegated rotation event.
-    let stream = r#"{"v":"KERI10JSON000122_","i":"E-9tsnVcfUyXVQyBPGfntoL-xexf4Cldt_EPzHis2W4U","s":"1","t":"drt","p":"E1x1JOub6oEQkxAxTNFu1Pma6y-lrbprNsaILHJHoPmY","kt":"1","k":["DTf6QZWoet154o9wvzeMuNhLQRr8JaAUeiC6wjB_4_08"],"n":"E8kyiXDfkE7idwWnAZQjHbUZMz-kd_yIMH0miptIFFPo","bt":"0","br":[],"ba":[],"a":[]}"#.as_bytes();
-    let event = message(stream);
-    assert!(event.is_ok());
-    assert_eq!(event.unwrap().1.serialize().unwrap(), stream);
-}
-
-#[test]
-fn test_signed_events_stream() {
-    let kerl_str = br#"{"v":"KERI10JSON0000ed_","i":"DoQy7bwiYr80qXoISsMdGvfXmCCpZ9PUqetbR8e-fyTk","s":"0","t":"icp","kt":"1","k":["DoQy7bwiYr80qXoISsMdGvfXmCCpZ9PUqetbR8e-fyTk"],"n":"EGofBtQtAeDMOO3AA4QM0OHxKyGQQ1l2HzBOtrKDnD-o","bt":"0","b":[],"c":[],"a":[]}-AABAAxemWo-mppcRkiGSOXpVwh8CYeTSEJ-a0HDrCkE-TKJ-_76GX-iD7s4sbZ7j5fdfvOuTNyuFw3a797gwpnJ-NAg{"v":"KERI10JSON000122_","i":"DoQy7bwiYr80qXoISsMdGvfXmCCpZ9PUqetbR8e-fyTk","s":"1","t":"rot","p":"EvZY9w3fS1h98tJeysdNQqT70XLLec4oso8kIYjfu2Ks","kt":"1","k":["DLqde_jCw-C3y0fTvXMXX5W7QB0188bMvXVkRcedgTwY"],"n":"EW5MfLjWGOUCIV1tQLKNBu_WFifVK7ksthNDoHP89oOc","bt":"0","br":[],"ba":[],"a":[]}-AABAAuQcoYU04XYzJxOPp4cxmvXbqVpGADfQWqPOzo1S6MajUl1sEWEL1Ry30jNXaV3-izvHRNROYtPm2LIuIimIFDg{"v":"KERI10JSON000122_","i":"DoQy7bwiYr80qXoISsMdGvfXmCCpZ9PUqetbR8e-fyTk","s":"2","t":"rot","p":"EOi_KYKjP4hinuTfgtoYj5QBw_Q1ZrRtWFQDp0qsNuks","kt":"1","k":["De5pKs8wiP9bplyjspW9L62PEANoad-5Kum1uAllRxPY"],"n":"ERKagV0hID1gqZceLsOV3s7MjcoRmCaps2bPBHvVQPEQ","bt":"0","br":[],"ba":[],"a":[]}-AABAAPKIYNAm6nmz4cv37nvn5XMKRVzfKkVpJwMDt2DG-DqTJRCP8ehCeyDFJTdtvdJHjKqrnxE4Lfpll3iUzuQM4Aw{"v":"KERI10JSON000122_","i":"DoQy7bwiYr80qXoISsMdGvfXmCCpZ9PUqetbR8e-fyTk","s":"3","t":"rot","p":"EVK1FbLl7yWTxOzPwk7vo_pQG5AumFoeSE51KapaEymc","kt":"1","k":["D2M5V_e23Pa0IAqqhNDKzZX0kRIMkJyW8_M-gT_Kw9sc"],"n":"EYJkIfnCYcMFVIEi-hMMIjBQfXcTqH_lGIIqMw4LaeOE","bt":"0","br":[],"ba":[],"a":[]}-AABAAsrKFTSuA6tEzqV0C7fEbeiERLdZpStZMCTvgDvzNMfa_Tn26ejFRZ_rDmovoo8xh0dH7SdMQ5B_FvwCx9E98Aw{"v":"KERI10JSON000098_","i":"DoQy7bwiYr80qXoISsMdGvfXmCCpZ9PUqetbR8e-fyTk","s":"4","t":"ixn","p":"EY7VDg-9Gixr9rgH2VyWGvnnoebgTyT9oieHZIaiv2UA","a":[]}-AABAAqHtncya5PNnwSbMRegftJc1y8E4tMZwajVVj2-FmGmp82b2A7pY1vr7cv36m7wPRV5Dusf4BRa5moMlHUpSqDA"#;
-    let (rest, messages) = signed_event_stream(kerl_str).unwrap();
-
-    assert!(rest.is_empty());
-    assert_eq!(messages.len(), 5);
+    // Nontrans receipt with alternative attachment with -B payload type. Not implemented yet.
+    // let witness_receipts = r#"{"v":"KERI10JSON000091_","i":"EpU9D_puIW_QhgOf3WKUy-gXQnXeTQcJCO_Igcxi1YBg","s":"0","t":"rct","d":"EIt0xQQf-o-9E1B9VTDHiicQzVWk1CptvnewcnuhSd0M"}-BADAACZrTPLvG7sNaxtV8ZGdIHABFHCZ9FlnG6b4J6a9GcyzJIJOjuGNphW2zyC_WWU6CGMG7V52UeJxPqLpaYdP7CgAB8npsG58rX1ex73gaGe-jvRnw58RQGsDLzoSXaGn-kHRRNu6Kb44zXDtMnx-_8CjnHqskvDbz6pbEbed3JTOnCQACM4bMcLjcDtD0fmLOGDx2oxBloc2FujbyllA7GuPLm-RQbyPPQr70_Y7DXzlWgs8gaYotUATeR-dj1ru9qFwADA"#;
+    // let msg = signed_message(witness_receipts.as_bytes());
+    // assert!(msg.is_ok());
 }
